requirement.txt
certifi==2019.11.28
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.2
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
pandas==0.25.3
Pillow==7.0.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
torch==1.4.0
torchvision==0.5.0
